CLH: Stuff being trained

+++ GA3C-master/ga3c/ProcessAgent.py
@@ -64,7 +64,7 @@ class ProcessAgent(Process):
 
     def convert_data(self, experiences):
         x_ = np.array([exp.state for exp in experiences])
-        a_ = np.eye(self.num_actions)[np.array([exp.action for exp in experiences])].astype(np.float32)
+        a_ = np.eye(self.num_actions)[np.array([exp.action for exp in experiences]).astype(int)].astype(np.float32)
         r_ = np.array([exp.reward for exp in experiences])
         return x_, r_, a_
